Move polyfill classes and enums into their own files; Style fixes
