+++ Scripts/Core/EditorVR.cs
@@ -13,44 +13,6 @@ using UnityEngine.InputNew;
 [assembly: OptionalDependency("PolyToolkit.PolyApi", "INCLUDE_POLY_TOOLKIT")]
 [assembly: OptionalDependency("UnityEngine.DrivenRectTransformTracker+BlockUndoCCU", "UNDO_PATCH")]
 
-#if !UNITY_EDITOR
-        class EditorPrefs
-        {
-            public static bool GetBool(string key, bool defaultValue)
-            {
-                var value = PlayerPrefs.GetString(key, defaultValue.ToString());
-                bool result;
-                bool.TryParse(value, out result);
-                return result;
-            }
-
-            public static void SetBool(string key, bool value)
-            {
-                PlayerPrefs.SetString(key, value.ToString());
-            }
-
-            public static string GetString(string key, string defaultValue)
-            {
-                return PlayerPrefs.GetString(key, defaultValue);
-            }
-
-            public static string GetString(string key)
-            {
-                return PlayerPrefs.GetString(key);
-            }
-
-            public static void SetString(string key, string value)
-            {
-                PlayerPrefs.SetString(key, value);
-            }
-
-            public static bool HasKey(string key)
-            {
-                return PlayerPrefs.HasKey(key);
-            }
-        }
-#endif
-
 namespace UnityEditor.Experimental.EditorVR.Core
 {
 #if UNITY_2017_2_OR_NEWER
