+++ Scripts/Interfaces/FunctionalityInjection/IUsesSnapping.cs
@@ -1,14 +1,6 @@
 ﻿using System;
 using UnityEngine;
 
-#if !UNITY_EDITOR
-public enum PivotMode
-{
-    Center,
-    Pivot,
-}
-#endif
-
 namespace UnityEditor.Experimental.EditorVR
 {
     /// <summary>
