+++ Scripts/Runtime/EditorPrefs.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: ecab5ed034c4a9e4f8d2ce47dcdbe842
+timeCreated: 1481295560
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
