+++ Scripts/Runtime/PivotMode.cs
@@ -1,0 +1,10 @@
+﻿#if !UNITY_EDITOR
+namespace UnityEditor.Experimental.EditorVR
+{
+    public enum PivotMode
+    {
+        Center,
+        Pivot,
+    }
+}
+#endif
