+++ Scripts/Runtime/PivotMode.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: ce3c22c246369df4798a8835c5b4cb38
+timeCreated: 1481295560
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
