+++ Scripts/Runtime/SerializedObject.cs
@@ -1,0 +1,11 @@
+﻿#if !UNITY_EDITOR
+using UnityEngine;
+
+namespace UnityEditor.Experimental.EditorVR
+{
+    public class SerializedObject
+    {
+        public GameObject targetObject;
+    }
+}
+#endif
