+++ Scripts/Runtime/SerializedObject.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 81a95100795a22e438d30eb1abe6b466
+timeCreated: 1481295560
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
