+++ Scripts/Runtime/SerializedPropertyType.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 3ad2c32e80d3f8f4eb657d93918fbce5
+timeCreated: 1481295560
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
